requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
pyserial>=3.5
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
svgwrite>=1.4.3
svgpathtools>=1.6.0
cairosvg>=2.7.0
gunicorn>=21.0.0
eventlet>=0.33.0
python-engineio>=4.8.0
python-socketio>=5.10.0

